Fix building with Xcode 12.5

+++ Source/OpenEmuHelperApp.swift
@@ -415,7 +415,7 @@ extension OSLog {
         
         if let parameters = parameters, let filter = _filterChain {
             for (key, val) in parameters {
-                filter.setValue(val, forParameterName: key)
+                filter.setValue(CGFloat(val), forParameterName: key)
             }
         }
     }
